feat: separate CardDetailModal component
